# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenWrapper Gateway API",
    "version": "0.1.4",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
...
  },
  "servers": [
    {
      "url": "https://gateway.openwrapper.muejam.com",
      "description": "Production OpenWrapper Gateway (TLS - Paymob, Fawry, Stripe)"
    },
    {
      "url": "https://openwrapper.muejam.com/api",
      "description": "Production Web API (Control Plane, Better Auth, & Ledger GraphQL)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Rust gateway (Paymob, Fawry, Stripe, gRPC & GraphQL)"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Web Control Plane & API"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/webhooks/{provider}": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Ingest a provider webhook",
        "operationId": "ingestWebhook",
        "security": [],
        "description": "Provider-facing callback endpoint. It is authenticated using the\nprovider's signature rather than a gateway API key. Paymob sends its\nHMAC-SHA512 value in the `hmac` query parameter, Fawry includes\n`messageSignature` in its JSON payload, and Stripe sends a timestamped\nHMAC-SHA256 signature in the `Stripe-Signature` header (`t=...,v1=...`).\n\nBoth the Rust gateway and web API support Paymob, Fawry, and Stripe natively.\nSuccessful responses may have an empty body; callers must use the HTTP\nstatus as the acknowledgement.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/WebhookProvider"
          },
          {
            "name": "hmac",
            "in": "query",
            "required": false,
            "description": "Paymob HMAC-SHA512 signature.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Stripe-Signature",
            "in": "header",
            "required": false,
            "description": "Stripe webhook signature header (`t=...,v1=...`) with timestamp replay tolerance check.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Raw provider-owned webhook payload.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderWebhookPayload"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ProviderWebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verified and acknowledged, including duplicate deliveries."
          },
          "400": {
            "description": "Malformed payload, unsupported provider, or failed consistency validation."
          },
          "401": {
            "description": "Missing or invalid provider signature."
          },
          "404": {
            "description": "Unknown provider."
          },
          "500": {
            "description": "Webhook could not be durably queued or applied."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "OpenWrapper API key in `Authorization: Bearer <key>`."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "OpenWrapper API key sent directly in `X-API-Key`."
      }
    },
    "parameters": {
...
      "WebhookProvider": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/PaymentProvider"
        }
      },
...
    },
...
    "schemas": {
      "PaymentProvider": {
        "type": "string",
        "enum": [
          "paymob",
          "fawry",
          "stripe",
          "mock"
        ],
        "description": "Provider adapter. Native first-class support for `paymob`, `fawry`,\n`stripe`, and `mock` across both the Rust gateway and web API.\n"
      },
...
      "ProviderWebhookPayload": {
        "type": "object",
        "additionalProperties": true,
        "description": "Provider-owned payload retained as an open object because each provider signs a different wire model."
      },
...
    }
  }
}